refactor: move <MapEvents /> to useMapEvents hook
